test: migrate zero-integrations-telegram-message to bdd
